+++ src/Schema/parse/traverseSchema.js
@@ -1,34 +1,33 @@
 import { appendToPath, lastInPath, traverse } from 'object-agent';
-import { is } from 'type-enforcer';
-import Model from '../Model';
+import { isArray, isNumber, isObject } from 'type-enforcer';
 import checkSchemaType, { isType } from './checkSchemaType';
 
-export default function(item, callback) {
+export default (item, callback) => {
 	const subTraverse = (basePath, item) => {
 		traverse(item, (path, value) => {
 			const last = lastInPath(path);
 
-			if (is.number(last, true) && last > 0) {
+			if (isNumber(last, true) && last > 0) {
 				return true;
 			}
 
-			const isAnObject = is.object(value);
+			const isAnObject = isObject(value);
 			const isSchemaType = checkSchemaType(value);
 
 			if (!isSchemaType) {
-				value = isAnObject ? Object : is.array(value) ? Array : value;
-			}
-
-			if (is.instanceOf(value, Model)) {
-				subTraverse(appendToPath(basePath, path), value.schema);
-				return true;
+				if (isAnObject) {
+					value = Object;
+				}
+				else if (isArray(value)) {
+					value = Array;
+				}
 			}
 
 			if (basePath === '' || path !== '') {
 				callback(appendToPath(basePath, path), value, isAnObject, isSchemaType);
 			}
 
-			if (isSchemaType && isAnObject && ('content' in value)) {
+			if (isSchemaType && isAnObject && value.content !== undefined) {
 				subTraverse(appendToPath(basePath, path), isType(value.content) ? [[value.content]] : [value.content]);
 			}