+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## unreleased
+### Changed
+- [Schema](docs/Schema.md)
+  - Improved support for nested schemas
+  - Added support for '*' key
+
 ## [0.4.1] - 2019-07-01
 ### Added
 - Method 'extend' in [Schema](docs/Schema.md)